# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentAnswerOut": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "basis": {
            "items": {
              "$ref": "#/components/schemas/AgentRefOut"
            },
            "title": "Basis",
            "type": "array"
          },
          "caveat": {
            "title": "Caveat",
            "type": "string"
          },
          "computed_at": {
            "title": "Computed At",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "key",
          "question",
          "answer",
          "basis",
          "caveat",
          "computed_at"
        ],
        "title": "AgentAnswerOut",
        "type": "object"
      },
      "AgentContextOut": {
        "additionalProperties": true,
        "description": "Free-form by design: the shape evolves with the engines. Managers only.",
        "properties": {
          "caveat": {
            "title": "Caveat",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "today": {
            "title": "Today",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "today",
          "caveat"
        ],
        "title": "AgentContextOut",
        "type": "object"
      },
      "AgentQuestionOut": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "key",
          "question"
        ],
        "title": "AgentQuestionOut",
        "type": "object"
      },
      "AgentRefOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id",
          "title"
        ],
        "title": "AgentRefOut",
        "type": "object"
      },
...
      "CustomerType": {
        "enum": [
          "b2b",
          "b2c",
          "ambos"
        ],
        "title": "CustomerType",
        "type": "string"
      },
...
      "DecisionOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "why": {
            "title": "Why",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id",
          "title",
          "why"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
...
      "ExecutiveSummaryOut": {
        "properties": {
          "caveat": {
            "title": "Caveat",
            "type": "string"
          },
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "controls": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Controls",
            "type": "object"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionOut"
            },
            "title": "Decisions",
            "type": "array"
          },
          "documents": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Documents",
            "type": "object"
          },
          "exposures": {
            "items": {
              "$ref": "#/components/schemas/ExposureOut"
            },
            "title": "Exposures",
            "type": "array"
          },
          "improved": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Improved",
            "type": "object"
          },
          "next_30_days": {
            "items": {
              "$ref": "#/components/schemas/PriorityItemOut"
            },
            "title": "Next 30 Days",
            "type": "array"
          },
          "score": {
            "additionalProperties": true,
            "title": "Score",
            "type": "object"
          },
          "trend": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrendOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "unplanned": {
            "items": {
              "$ref": "#/components/schemas/UnplannedRiskOut"
            },
            "title": "Unplanned",
            "type": "array"
          },
          "window_days": {
            "title": "Window Days",
            "type": "integer"
          },
          "worsened": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Worsened",
            "type": "object"
          }
        },
        "required": [
          "computed_at",
          "window_days",
          "score",
          "trend",
          "exposures",
          "improved",
          "worsened",
          "decisions",
          "next_30_days",
          "unplanned",
          "controls",
          "documents",
          "caveat"
        ],
        "title": "ExecutiveSummaryOut",
        "type": "object"
      },
      "ExposureOut": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/RiskCategory"
          },
          "coverage": {
            "title": "Coverage",
            "type": "number"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "planned": {
            "title": "Planned",
            "type": "boolean"
          },
          "risk_id": {
            "format": "uuid",
            "title": "Risk Id",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/RiskSeverity"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "risk_id",
          "title",
          "severity",
          "category",
          "planned",
          "coverage",
          "due_date"
        ],
        "title": "ExposureOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeadcountBand": {
        "enum": [
          "ate_9",
          "de_10_a_49",
          "de_50_a_199",
          "acima_de_200"
        ],
        "title": "HeadcountBand",
        "type": "string"
      },
...
      "PrioritiesOut": {
        "properties": {
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "current_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Score"
          },
          "due_soon_days": {
            "title": "Due Soon Days",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PriorityItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "profile_complete": {
            "title": "Profile Complete",
            "type": "boolean"
          },
          "total_pending": {
            "title": "Total Pending",
            "type": "integer"
          },
          "unplanned": {
            "items": {
              "$ref": "#/components/schemas/UnplannedRiskOut"
            },
            "title": "Unplanned",
            "type": "array"
          }
        },
        "required": [
          "computed_at",
          "current_score",
          "profile_complete",
          "items",
          "total_pending",
          "unplanned",
          "due_soon_days"
        ],
        "title": "PrioritiesOut",
        "type": "object"
      },
      "PriorityItemOut": {
        "properties": {
          "action_id": {
            "format": "uuid",
            "title": "Action Id",
            "type": "string"
          },
          "control_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Control Id"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "effort": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionEffort"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "points": {
            "title": "Points",
            "type": "number"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "risk_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Id"
          },
          "risk_severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskSeverity"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Title"
          },
          "score_gain": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Gain"
          },
          "status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "action_id",
          "title",
          "status",
          "due_date",
          "risk_id",
          "risk_title",
          "risk_severity",
          "control_id",
          "effort",
          "points",
          "score_gain",
          "reasons"
        ],
        "title": "PriorityItemOut",
        "type": "object"
      },
      "ProfileOut": {
        "properties": {
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "customer_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerType"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_categories": {
            "items": {
              "type": "string"
            },
            "title": "Data Categories",
            "type": "array"
          },
          "headcount_band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HeadcountBand"
              },
              {
                "type": "null"
              }
            ]
          },
          "international_transfers": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Tristate"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "processes": {
            "items": {
              "type": "string"
            },
            "title": "Processes",
            "type": "array"
          },
          "segment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Segment"
              },
              {
                "type": "null"
              }
            ]
          },
          "sells_to_enterprise": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sells To Enterprise"
          },
          "systems": {
            "items": {
              "type": "string"
            },
            "title": "Systems",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "segment",
          "headcount_band",
          "customer_type",
          "data_categories",
          "sells_to_enterprise",
          "international_transfers",
          "systems",
          "processes",
          "notes",
          "complete",
          "updated_at"
        ],
        "title": "ProfileOut",
        "type": "object"
      },
      "ProfileUpdate": {
        "additionalProperties": false,
        "description": "Partial update; every field optional so the form can be filled in steps.",
        "properties": {
          "customer_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerType"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_categories": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DataCategory"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Categories"
          },
          "headcount_band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HeadcountBand"
              },
              {
                "type": "null"
              }
            ]
          },
          "international_transfers": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Tristate"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "processes": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processes"
          },
          "segment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Segment"
              },
              {
                "type": "null"
              }
            ]
          },
          "sells_to_enterprise": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sells To Enterprise"
          },
          "systems": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Systems"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "RadarOut": {
        "properties": {
          "all_clear": {
            "title": "All Clear",
            "type": "boolean"
          },
          "computed_at": {
            "format": "date",
            "title": "Computed At",
            "type": "string"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RadarItemOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "computed_at",
          "items",
          "counts",
          "all_clear"
        ],
        "title": "RadarOut",
        "type": "object"
      },
...
      "RiskCategory": {
        "enum": [
          "dados",
          "acesso",
          "seguranca",
          "fornecedores",
          "documentacao",
          "titulares",
          "incidentes",
          "pessoas"
        ],
        "title": "RiskCategory",
        "type": "string"
      },
...
      "Segment": {
        "enum": [
          "software_saas",
          "servicos",
          "comercio",
          "industria",
          "saude",
          "educacao",
          "financeiro",
          "outro"
        ],
        "title": "Segment",
        "type": "string"
      },
...
      "TrendOut": {
        "properties": {
          "diff": {
            "title": "Diff",
            "type": "integer"
          },
          "from_": {
            "title": "From",
            "type": "integer"
          },
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          },
          "to": {
            "title": "To",
            "type": "integer"
          }
        },
        "required": [
          "from_",
          "to",
          "diff",
          "since"
        ],
        "title": "TrendOut",
        "type": "object"
      },
      "Tristate": {
        "enum": [
          "sim",
          "nao",
          "nao_sei"
        ],
        "title": "Tristate",
        "type": "string"
      },
      "UnplannedRiskOut": {
        "properties": {
          "has_control": {
            "title": "Has Control",
            "type": "boolean"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_id": {
            "format": "uuid",
            "title": "Risk Id",
            "type": "string"
          },
          "score_gain": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Gain"
          },
          "severity": {
            "$ref": "#/components/schemas/RiskSeverity"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "risk_id",
          "title",
          "severity",
          "has_control",
          "score_gain"
        ],
        "title": "UnplannedRiskOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Compliance OS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/orgs/{org_id}/agent/answers/{key}": {
      "get": {
        "operationId": "agent_answer_api_v1_orgs__org_id__agent_answers__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAnswerOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Answer",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/agent/context": {
      "get": {
        "operationId": "agent_context_api_v1_orgs__org_id__agent_context_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentContextOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Context",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/agent/questions": {
      "get": {
        "operationId": "agent_questions_api_v1_orgs__org_id__agent_questions_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentQuestionOut"
                  },
                  "title": "Response Agent Questions Api V1 Orgs  Org Id  Agent Questions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Questions",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/api/v1/orgs/{org_id}/executive-summary": {
      "get": {
        "operationId": "get_executive_summary_api_v1_orgs__org_id__executive_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutiveSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Executive Summary",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/api/v1/orgs/{org_id}/priorities": {
      "get": {
        "operationId": "get_priorities_api_v1_orgs__org_id__priorities_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrioritiesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Priorities",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/profile": {
      "get": {
        "operationId": "get_profile_api_v1_orgs__org_id__profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "operationId": "update_profile_api_v1_orgs__org_id__profile_put",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/v1/orgs/{org_id}/radar": {
      "get": {
        "operationId": "get_radar_api_v1_orgs__org_id__radar_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RadarOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Radar",
        "tags": [
          "insights"
        ]
      }
    },
...
  }
}